add visual test to cover package
